requirements/constraints.txt
# Version constraints for pip-installation.
#
# This file doesn't install any packages. It specifies version constraints
# that will be applied if a package is needed.
#
# When pinning something here, please provide an explanation of why.  Ideally,
# link to other information that will help people in the future to remove the
# pin when possible.  Writing an issue against the offending project and
# linking to it here is good.

# TODO: Add constraint explanation

# edx-platform constraints
# be aware of the platform version used
-c https://raw.githubusercontent.com/openedx/edx-platform/open-release/redwood.master/requirements/constraints.txt


XBlock==4.0.1
beautifulsoup4==4.12.3
celery==5.4.0
django-config-models==2.7.0
django-countries==7.6.1
edx-drf-extensions==10.3.0
edx-opaque-keys==2.9.0
edxval==2.5.0
fs==2.0.27
jsonfield==3.1.0
openedx-events==9.10.0
path==16.11.0
pip==23.2.1
pip-tools==7.3.0
pylint==2.15.10
pycodestyle==2.8.0
requests==2.31.0
